refactor trace console e2e tests
